requirements.txt
babel==2.18.0
blinker==1.9.0
certifi==2026.1.4
charset-normalizer==3.4.4
click==8.3.1
courlan==1.3.2
dateparser==1.2.2
Flask==3.1.2
google_search_results==2.4.2
greenlet==3.3.1
htmldate==1.9.4
idna==3.11
itsdangerous==2.2.0
Jinja2==3.1.6
jusText==3.0.2
lxml==6.0.2
lxml_html_clean==0.4.3
MarkupSafe==3.0.3
PyMySQL==1.1.2
python-dateutil==2.9.0.post0
python-docx==1.2.0
python-dotenv==1.2.1
pytz==2025.2
regex==2026.1.15
requests==2.32.5
six==1.17.0
SQLAlchemy==2.0.46
tld==0.13.1
trafilatura==2.0.0
typing_extensions==4.15.0
tzlocal==5.3.1
urllib3==2.6.3
Werkzeug==3.1.5
gunicorn
